pyproject.toml
[tool.poetry]
name = "yfbot"
version = "0.1.0"
description = ""
authors = ["narumi <toucans-cutouts0f@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.11"
yfinance = "^0.2.40"
python-dotenv = "^1.0.1"
python-telegram-bot = "^21.2"
loguru = "^0.7.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-cov = "^5.0.0"
ruff = "^0.4.7"
toml = "^0.10.2"
mypy = "^1.10.0"

[tool.poetry.scripts]
yfbot = "yfbot.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = ["build"]
line-length = 120

[tool.ruff.lint]
select = [
    "B",   # flake8-bugbear
    "C",   # flake8-comprehensions
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "SIM", # flake8-simplify
    "UP",  # pyupgrade
    "W",   # pycodestyle warnings
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]

[tool.ruff.lint.isort]
force-single-line = true

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning"]
